Refactor: Use 2-Dimensional Int Array instead of 1-Dimensional Tile Array
(kinda)Big commit… mainly necessary (except for the renamed of ‘tiles’
to ‘grid’)

Bottleneck should be filtering out a single Tile, row * column times in
a row * column array, when the board gets large this is about 75%+ of
processing time based on internal testing of getUniqueTile -> Not
committed as it is a private function and it’s not meant to be public.

Since the row and column are now stored as an the index in the
array(albeit row/column  -  1), the Tile class simply holds an Int
Value, so now just a 2-dimensional array of Ints, So remove the Tile
class

Huge increase in performance, small board 17% better (still 0.001
though)
medium board 67%, and huge board 89% (0.384s -> 0.042s).

I know you guys at pillar probably won’t see this but it was bothering
me anyways, it needed fixed!